pyproject.toml
[project]
name = "dashfuel-takehome"
version = "0.1.0"
description = ""
authors = [
    {name = "Clark Burns",email = "rclarkburns@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
